+++ dbt_project/models/staging/stg_liquidity_premium.sql
@@ -1,0 +1,8 @@
+{{ config(materialized='view') }}
+
+SELECT
+    valuation_date,
+    tenor_label,
+    tenor_years,
+    lp_bps
+FROM {{ source('raw', 'liquidity_premium') }}