+++ dbt_project/models/staging/stg_risk_ftp_curve.sql
@@ -1,0 +1,3 @@
+{{ config(materialized='view') }}
+
+SELECT * FROM {{ source('risk', 'risk_ftp_curve') }}