+++ dbt_project/models/staging/stg_risk_nii_attribution_rows.sql
@@ -1,0 +1,3 @@
+{{ config(materialized='view') }}
+
+SELECT * FROM {{ source('risk', 'risk_nii_attribution_rows') }}